+++ Magic/src/main/java/com/elmakers/mine/bukkit/citizens/MagicCitizensTrait.java
@@ -15,6 +15,7 @@ public class MagicCitizensTrait extends CitizensTrait {
 
     private String spellKey;
     private boolean npcCaster = false;
+    private boolean targetPlayer = true;
     private YamlConfiguration parameters = null;
 
 	public MagicCitizensTrait() {
@@ -26,6 +27,7 @@ public class MagicCitizensTrait extends CitizensTrait {
         super.load(data);
         spellKey = data.getString("spell", null);
         npcCaster = data.getBoolean("caster", false);
+        targetPlayer = data.getBoolean("target_player", true);
         String parameterString = data.getString("parameters", null);
         parameters = new YamlConfiguration();
         if (parameterString != null && !parameterString.isEmpty()) {
@@ -49,6 +51,7 @@ public class MagicCitizensTrait extends CitizensTrait {
         super.save(data);
         data.setString("spell", spellKey);
         data.setBoolean("caster", npcCaster);
+        data.setBoolean("target_player", targetPlayer);
         String parameterString = parameters == null ? null : parameters.saveToString();
         data.setString("parameters", parameterString);
 	}
@@ -64,9 +67,11 @@ public class MagicCitizensTrait extends CitizensTrait {
             if (event.getNPC().isSpawned()) {
                 entity = event.getNPC().getEntity();
                 sender = null;
-                config = new MemoryConfiguration();
-                ConfigurationUtils.addConfigurations(config, parameters);
-                config.set("player", event.getClicker().getName());
+                if (targetPlayer) {
+                    config = new MemoryConfiguration();
+                    ConfigurationUtils.addConfigurations(config, parameters);
+                    config.set("player", event.getClicker().getName());
+                }
             }
         }
 
@@ -80,6 +85,9 @@ public class MagicCitizensTrait extends CitizensTrait {
         String spellDescription = spellKey == null ? (ChatColor.RED + "(None)") : (ChatColor.LIGHT_PURPLE + spellKey);
         sender.sendMessage(ChatColor.DARK_PURPLE + "Spell: " + spellDescription);
         String casterDescription = npcCaster ? (ChatColor.GRAY + "NPC") : (ChatColor.LIGHT_PURPLE + "Player");
+        if (targetPlayer) {
+            casterDescription = casterDescription + ChatColor.DARK_GRAY + " (" + ChatColor.YELLOW + "Auto-target player" + ChatColor.DARK_GRAY + ")";
+        }
         sender.sendMessage(ChatColor.DARK_PURPLE + "Caster: " + casterDescription);
         sender.sendMessage(ChatColor.DARK_PURPLE + "Parameters: ");
         describeParameters(sender);
@@ -154,6 +162,22 @@ public class MagicCitizensTrait extends CitizensTrait {
             {
                 npcCaster = true;
                 sender.sendMessage(ChatColor.DARK_PURPLE + "Set caster as NPC");
+            }
+        }
+        else if (key.equalsIgnoreCase("target_player"))
+        {
+            if (value == null || !value.equalsIgnoreCase("true"))
+            {
+                sender.sendMessage(ChatColor.DARK_PURPLE + "Will not auto-target the clicking player");
+                targetPlayer = false;
+            }
+            else
+            {
+                targetPlayer = true;
+                sender.sendMessage(ChatColor.DARK_PURPLE + "Will auto-target the clicking player");
+            }
+            if (!npcCaster) {
+                sender.sendMessage(ChatColor.RED + "NOTE: " + ChatColor.YELLOW + "Has no effect unless you also set " + ChatColor.AQUA + "caster true");
             }
         }
         else if (value == null || value.isEmpty())